// WADNR.Web/src/app/shared/generated/model/support-request-create.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class SupportRequestCreate { 
    SupportRequestTypeID: number;
    RequestDescription: string;
    RequestPersonOrganization?: string | null;
    RequestPersonPhone?: string | null;
    CurrentPageUrl?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface SupportRequestCreateForm { 
    SupportRequestTypeID: FormControl<number>;
    RequestDescription: FormControl<string>;
    RequestPersonOrganization?: FormControl<string>;
    RequestPersonPhone?: FormControl<string>;
    CurrentPageUrl?: FormControl<string>;
}

export class SupportRequestCreateFormControls { 
    public static SupportRequestTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
            ],
        }
    );
    public static RequestDescription = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
                Validators.minLength(0),
                Validators.maxLength(2000),
            ],
        }
    );
    public static RequestPersonOrganization = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
                Validators.minLength(0),
                Validators.maxLength(500),
            ],
        }
    );
    public static RequestPersonPhone = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
                Validators.minLength(0),
                Validators.maxLength(50),
            ],
        }
    );
    public static CurrentPageUrl = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
                Validators.minLength(0),
                Validators.maxLength(500),
            ],
        }
    );
}